TCIITensorConversion

Documentation for TCIITensorConversion.

ITensorMPS.MPS

TensorCrossInterpolation.TensorTrain

TensorCrossInterpolation.TensorTrain

TCIITensorConversion.evaluate_mps

TCIITensorConversion.evaluate_mps

ITensorMPS.MPS — Method

MPS(tt, siteindices...)

Convert a tensor train to an ITensor MPS

tt Tensor train

siteindices Arrays of ITensor Index objects.

If siteindices is left empty, a default set of indices will be used.

TensorCrossInterpolation.TensorTrain — Method

function TCI.TensorTrain(mps::ITensorMPS.MPS)

Converts an ITensor MPS object into a TensorTrain. Note that this only works if the MPS has a single leg per site! Otherwise, use TCI.TensorTrain(mps::ITensorMPS.MPO).

TensorCrossInterpolation.TensorTrain — Method

function TCI.TensorTrain(mps::ITensorMPS.MPO)

Convertes an ITensor MPO object into a TensorTrain.

TCIITensorConversion.evaluate_mps — Method

function evaluate_mps(
    mps::Union{ITensorMPS.MPS,ITensorMPS.MPO},
    indices::AbstractVector{<:ITensorMPS.Index},
    indexvalues::AbstractVector{Int}
)

Evaluates an MPS or MPO for a given set of index values.

indices is a list of ITensorMPS.Index objects that specifies the order in which indices are given.

indexvalues is a list of integer values in the same order as indices.

If many evaluations are necessary, it may be advantageous to convert your MPS to a TensorCrossInterpolation.TTCache object first.

TCIITensorConversion.evaluate_mps — Method

function evaluate_mps(
    mps::Union{ITensorMPS.MPS,ITensorMPS.MPO},
    indexspecs::Vararg{AbstractVector{<:Tuple{ITensorMPS.Index,Int}}}
)

Evaluates an MPS or MPO for a given set of index values.

indexspec is a list of tuples, where each tuple contains an ITensorMPS.Index object specifying an index, and an Int corresponding to the value of the specified index.

If many evaluations are necessary, it may be advantageous to convert your MPS to a TensorCrossInterpolation.TTCache object first.